+++ apps/nextjs-website/src/components/molecules/DesktopSiteHeader/DesktopSiteHeader.tsx
@@ -5,8 +5,11 @@ import DesktopUserInfo from '@/components/atoms/DesktopUserInfo/DesktopUserInfo'
 import React from 'react';
 import { SiteHeaderProps } from '@/components/molecules/SiteHeader/SiteHeader';
 import { useTranslations } from 'next-intl';
+import LanguageSelector from '@/components/atoms/LanguageSelector/LanguageSelector';
+import { i18nActive } from '@/config';
+import { SUPPORTED_LOCALES } from '../../../locales';
 
-const DesktopSiteHeader = ({ products }: SiteHeaderProps) => {
+const DesktopSiteHeader = ({ locale, products }: SiteHeaderProps) => {
   const t = useTranslations('devPortal');
 
   return (
@@ -26,6 +29,11 @@ const DesktopSiteHeader = ({ products }: SiteHeaderProps) => {
           href: `/${product.slug}/overview`,
           label: product.name,
         }))}
+        menuStyle={{
+          sx: {
+            marginTop: '17px',
+          },
+        }}
       />
       <LinkMui
         component={Link}
@@ -46,6 +54,9 @@ const DesktopSiteHeader = ({ products }: SiteHeaderProps) => {
         {t('siteHeader.webinars')}
       </LinkMui>
       <DesktopUserInfo />
+      {i18nActive && (
+        <LanguageSelector locales={SUPPORTED_LOCALES} currentLocale={locale} />
+      )}
     </Box>
   );
 };